+++ e2e/reports/intergration-report/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @coinmasters/integration-coins
 
+## 1.13.23
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.26
+
+## 1.13.22
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.25
+
+## 1.13.21
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.24
+
 ## 1.13.20
 
 ### Patch Changes