+++ e2e/wallets/eth-asset-context/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @coinmasters/eth-asset-context-test
 
+## 1.0.23
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.26
+
+## 1.0.22
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.25
+
+## 1.0.21
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.24
+
 ## 1.0.20
 
 ### Patch Changes