+++ packages/pioneer/pioneer-react/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @coinmasters/pioneer-react
 
+## 0.10.5
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.26
+
+## 0.10.4
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.25
+
+## 0.10.3
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.24
+
 ## 0.10.2
 
 ### Patch Changes