+++ packages/pioneer/pioneer-sdk/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @coinmasters/pioneer-sdk
 
+## 4.13.26
+
+### Patch Changes
+
+- bump
+
+## 4.13.25
+
+### Patch Changes
+
+- bump
+
+## 4.13.24
+
+### Patch Changes
+
+- bump
+
 ## 4.13.23
 
 ### Patch Changes